Fix endless gulp loop...
The gulp 'copyrss.js' task was copying the rss file from the '_site' dir to the jekyll root.
The gulp watch task would then see this as a modification and rebuild the site.
At the end of the site build the copyrss task is run again starting the loop...

**The solution:**

Delete 'copyrss.js'
Th gulp copyrss task is no longer needed as the rss can be read directly from Web03.