Add the commented out styles for the `.header` class into the startup...
These styles control the main graphic on the top of landing pages i.e. the background-image for .header.
Because that background graphic is unique in each landing page, the styles should be easy to get at when making a new landing page.
Just uncomment and adjust where the `background-image` properties value.